#8 Move validation checkers to validation.utils
